remove unnecessary float32 casting

+++ cluster-autoscaler/cloudprovider/clusterapi/clusterapi_provider.go
@@ -158,7 +158,7 @@ func BuildClusterAPI(opts config.AutoscalingOptions, do cloudprovider.NodeGroupD
 	if err != nil {
 		klog.Fatalf("cannot build management cluster config: %v", err)
 	}
-	managementConfig.QPS = float32(opts.KubeClientOpts.KubeClientQPS)
+	managementConfig.QPS = opts.KubeClientOpts.KubeClientQPS
 	managementConfig.Burst = opts.KubeClientOpts.KubeClientBurst
 
 	workloadKubeconfig := opts.KubeClientOpts.KubeConfigPath
@@ -167,7 +167,7 @@ func BuildClusterAPI(opts config.AutoscalingOptions, do cloudprovider.NodeGroupD
 	if err != nil {
 		klog.Fatalf("cannot build workload cluster config: %v", err)
 	}
-	workloadConfig.QPS = float32(opts.KubeClientOpts.KubeClientQPS)
+	workloadConfig.QPS = opts.KubeClientOpts.KubeClientQPS
 	workloadConfig.Burst = opts.KubeClientOpts.KubeClientBurst
 
 	// Grab a dynamic interface that we can create informers from
